changelog: Internal, reCaptcha optimization, Load reCAPTCHA script earlier in page
