+++ spec/components/captcha_submit_button_component_spec.rb
@@ -45,22 +45,6 @@ RSpec.describe CaptchaSubmitButtonComponent, type: :component do
         "lg-captcha-submit-button[recaptcha-site-key='#{recaptcha_site_key}']",
       )
     end
-
-    it 'renders script tag for recaptcha' do
-      src = "https://www.google.com/recaptcha/api.js?render=#{recaptcha_site_key}"
-      expect(rendered).to have_css("script[src='#{src}']", visible: :all)
-    end
-
-    context 'with recaptcha enterprise' do
-      before do
-        allow(FeatureManagement).to receive(:recaptcha_enterprise?).and_return(true)
-      end
-
-      it 'renders script tag for recaptcha' do
-        src = "https://www.google.com/recaptcha/enterprise.js?render=#{recaptcha_site_key}"
-        expect(rendered).to have_css("script[src='#{src}']", visible: :all)
-      end
-    end
   end
 
   context 'with additional tag options' do
